Incub proverka

+++ app/src/main/java/com/zaroslikov/fermacompose2/data/ferma/Incubator.kt
@@ -25,5 +25,5 @@ data class Incubator(
     var over: String,
     var airing: String,
     @ColumnInfo(name = "idPT")
-    var idPT: Int
+    var idPT: Long
 )
